+++ lib/mongory/core_ext.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require 'mongory/core_ext/array'
-require 'mongory/core_ext/hash'
-require 'mongory/core_ext/symbol'
-require 'mongory/core_ext/string'
-require 'mongory/core_ext/object'
+require_relative 'core_ext/array'
+require_relative 'core_ext/hash'
+require_relative 'core_ext/symbol'
+require_relative 'core_ext/string'
+require_relative 'core_ext/object'