+++ lib/mongory/matchers/abstract_multi_matcher.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'mongory/matchers/abstract_matcher'
+require_relative 'abstract_matcher'
 
 module Mongory
   # Temp Description
@@ -8,18 +8,13 @@ module Mongory
     # Abstract class
     class AbstractMultiMatcher < AbstractMatcher
       def match?(record)
-        record = preprocess(record)
         matchers.send(operator) do |matcher|
           matcher.match?(record)
         end
       end
 
       define_instance_cache_method(:matchers) do
-        @condition.map(&method(:build_sub_matcher))
-      end
-
-      def preprocess(record)
-        record
+        @condition.map(&method(:build_sub_matcher)).uniq(&:condition)
       end
 
       def build_sub_matcher(*); end