+++ lib/mongory/matchers/key_value_matcher.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'mongory/matchers/abstract_matcher'
+require_relative 'abstract_matcher'
 
 module Mongory
   # Temp Description
@@ -27,21 +27,11 @@ module Mongory
       def match?(record)
         return operator_matcher.match?(record) if operator_matcher
 
-        sub_value = fetch_value(record)
-
-        if record.is_a?(Array) && sub_value == KEY_NOT_FOUND
-          elem_matcher.match?(record)
-        else
-          main_matcher.match?(sub_value)
-        end
+        main_matcher.match?(fetch_value(record))
       end
 
       define_matcher(:operator) do
         Matchers.lookup(@match_key)&.new(@match_value)
-      end
-
-      define_matcher(:elem) do
-        ElemMatchMatcher.new(@condition)
       end
 
       define_matcher(:main) do
@@ -55,9 +45,7 @@ module Mongory
         when Hash
           record.fetch(@match_key, KEY_NOT_FOUND)
         when Array
-          return KEY_NOT_FOUND unless @match_key.match?(/^\d+$/)
-
-          record[@match_key.to_i]
+          record[@match_key]
         when *CLASSES_NOT_ALLOW_TO_DIG
           KEY_NOT_FOUND
         else