+++ lib/mongory/matchers/lte_matcher.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'mongory/matchers/abstract_operator_matcher'
+require_relative 'abstract_operator_matcher'
 
 module Mongory
   # Temp Description