+++ lib/mongory/query_builder.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'mongory/utils'
+require_relative 'utils'
 
 module Mongory
   # Temp Description